Rename package cucumber_authentication to cucumberauthentication
