+++ src/main/java/tech/jhipster/lite/generator/server/springboot/cucumberauthentication/application/CucumberAuthenticationApplicationService.java
@@ -1,7 +1,7 @@
-package tech.jhipster.lite.generator.server.springboot.cucumber_authentication.application;
+package tech.jhipster.lite.generator.server.springboot.cucumberauthentication.application;
 
 import org.springframework.stereotype.Service;
-import tech.jhipster.lite.generator.server.springboot.cucumber_authentication.domain.CucumberAuthenticationModuleFactory;
+import tech.jhipster.lite.generator.server.springboot.cucumberauthentication.domain.CucumberAuthenticationModuleFactory;
 import tech.jhipster.lite.module.domain.JHipsterModule;
 import tech.jhipster.lite.module.domain.properties.JHipsterModuleProperties;
